Move table files to /src/Table
