--- packages/gestalt/src/Table.css
@@ -1,26 +1,0 @@
-.table {
-  border-collapse: separate;
-  border-spacing: 0;
-  width: 100%;
-}
-
-.th {
-  composes: alignLeft from "./Typography.css";
-  composes: paddingY3 from "./boxWhitespace.css";
-  composes: paddingX3 from "./boxWhitespace.css";
-}
-
-.td {
-  composes: paddingY3 from "./boxWhitespace.css";
-  composes: paddingX3 from "./boxWhitespace.css";
-}
-
-.sticky tr th {
-  background-color: #fff;
-  position: sticky;
-  top: 0;
-}
-
-.thead tr:last-child th {
-  border-bottom: 1px solid #ddd;
-}
