+++ packages/gestalt/src/Table/Table.css.flow
@@ -1,9 +1,9 @@
 // @flow strict
 /* This file is automatically generated by css-modules-flow-types */
 declare module.exports: {|
-  +'sticky': string,
-  +'table': string,
-  +'td': string,
-  +'th': string,
-  +'thead': string,
+  +sticky: string,
+  +table: string,
+  +td: string,
+  +th: string,
+  +thead: string,
 |};
