+++ packages/gestalt/src/Table/TableSortableHeaderCell.js
@@ -1,7 +1,9 @@
+/* eslint-disable eslint-comments/disable-enable-pair */
+/* eslint-disable import/no-relative-parent-imports */
 // @flow strict
 import * as React from 'react';
-import Box from './Box.js';
-import Icon from './Icon.js';
+import Box from '../Box.js';
+import Icon from '../Icon.js';
 import TableHeaderCell from './TableHeaderCell.js';
 import TapArea from './TapArea.js';
